refactor: gathering 생성 메서드 경로 수정 #138
- api/book-clubs/{bookClubId}/gatherings로 변경했습니다.
- 변경된 path에 따라서 GatheringController에서 BookClubController로 메서드를 이동했습니다.
- 변경된 path를 사용하도록 테스트를 변경했습니다.
- GatheringController과 GatheringTest를 다시 사용할 예정이라 삭제하지 않았습니다.
